minor aesthetic stuff

+++ py/cxx/pyhistogram.cxx
@@ -39,7 +39,7 @@
 
 static void check_pickle_tuple_size(std::size_t expect, int given)
 {
-  if (expect != given) {
+  if ((int)expect != given) {
     throw std::runtime_error(std::string("Invalid pickle state: expected ")+std::to_string(expect)+", got "
                              +std::to_string(given));
   }
